+++ client/out/extension.js
@@ -1,0 +1,45 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.deactivate = exports.activate = void 0;
+const path = require("path");
+const vscode_1 = require("vscode");
+const node_1 = require("vscode-languageclient/node");
+let client;
+function activate(context) {
+    const serverModule = context.asAbsolutePath(path.join("server", "out", "server.js"));
+    const debugOptions = { execArgv: ["--nolazy", "--inspect=6009"] };
+    const serverOptions = {
+        run: { module: serverModule, transport: node_1.TransportKind.ipc },
+        debug: {
+            module: serverModule,
+            transport: node_1.TransportKind.ipc,
+            options: debugOptions,
+        },
+    };
+    const clientOptions = {
+        documentSelector: [
+            { scheme: "file", language: "qml" },
+            { scheme: "untitle", language: "qml" },
+        ],
+        synchronize: {
+            // Notify the server about file changes to '.clientrc files contained in the workspace
+            fileEvents: [
+                vscode_1.workspace.createFileSystemWatcher("**/.qrc"),
+                vscode_1.workspace.createFileSystemWatcher("**/main.cpp"),
+            ],
+        },
+    };
+    // Create the language client and start the client.
+    client = new node_1.LanguageClient("qmlLanguageServer", "QML Language Server", serverOptions, clientOptions);
+    // Start the client. This will also launch the server
+    client.start();
+}
+exports.activate = activate;
+function deactivate() {
+    if (!client) {
+        return undefined;
+    }
+    return client.stop();
+}
+exports.deactivate = deactivate;
+//# sourceMappingURL=extension.js.map