+++ client/src/extension.ts
@@ -1,0 +1,59 @@
+import * as path from "path";
+import { workspace, ExtensionContext } from "vscode";
+
+import {
+  LanguageClient,
+  LanguageClientOptions,
+  ServerOptions,
+  TransportKind,
+} from "vscode-languageclient/node";
+
+let client: LanguageClient;
+
+export function activate(context: ExtensionContext) {
+  const serverModule = context.asAbsolutePath(
+    path.join("server", "out", "server.js")
+  );
+  const debugOptions = { execArgv: ["--nolazy", "--inspect=6009"] };
+
+  const serverOptions: ServerOptions = {
+    run: { module: serverModule, transport: TransportKind.ipc },
+    debug: {
+      module: serverModule,
+      transport: TransportKind.ipc,
+      options: debugOptions,
+    },
+  };
+
+  const clientOptions: LanguageClientOptions = {
+    documentSelector: [
+      { scheme: "file", language: "qml" },
+      { scheme: "untitle", language: "qml" },
+    ],
+    synchronize: {
+      // Notify the server about file changes to '.clientrc files contained in the workspace
+      fileEvents: [
+        workspace.createFileSystemWatcher("**/.qrc"),
+        workspace.createFileSystemWatcher("**/main.cpp"),
+      ],
+    },
+  };
+
+  // Create the language client and start the client.
+  client = new LanguageClient(
+    "qmlLanguageServer",
+    "QML Language Server",
+    serverOptions,
+    clientOptions
+  );
+
+  // Start the client. This will also launch the server
+  client.start();
+}
+
+export function deactivate(): Thenable<void> | undefined {
+  if (!client) {
+    return undefined;
+  }
+  return client.stop();
+}