+++ server/out/libs/regex.d.ts
@@ -1,0 +1,1 @@
+export declare function Regex(line: string, isfirstLine?: boolean, editor?: number): string;