+++ server/out/libs/stringFormatter.d.ts
@@ -1,0 +1,1 @@
+export declare function whiteSpaceRemove(str: string, spaceSize: number): string;